requirements.txt
streamlit==1.29.0
boto3==1.33.6
numpy==1.26.2
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
langchain==0.2.2
langchain-anthropic==0.1.16
langchain-core==0.2.10
langchain-google-genai==1.0.7
langchain-google-vertexai==1.0.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.82
huggingface_hub